api/serializers/tokens.py
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer


class MyTokenObtainPairSerializer(TokenObtainPairSerializer):
    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)

        # Add custom claims
        token['username'] = user.username
        token['email'] = user.email
        if user.role is not None:
            token['role'] = user.role.name
            token['is_staff'] = user.role.is_staff
        # ...

        return token
