requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
idna==3.4
marshmallow==3.19.0
packaging==23.0
pycodestyle==2.10.0
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
stripe==5.1.0
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
